+++ src/cpc/diskimage/extdsk.c
@@ -88,7 +88,7 @@ int		ExtDsk_Validate(const unsigned char *pDiskImage, const unsigned long DiskIm
 
 								thisTrackSize = 0;
 
-								pTrack = (EXTDSKTRACKHEADER *)((long)pHeader + CurrentSize + sizeof(EXTDSKTRACKHEADER));
+								pTrack = (EXTDSKTRACKHEADER *)((int)pHeader + CurrentSize + sizeof(EXTDSKTRACKHEADER));
 
 								/* check track header text is present */
 								if (memcmp(pTrack->TrackHeader,"Track-Info",10)!=0)
